refactor: Utils package 로 분류
