+++ src/test/java/utils/DoubleParserTest.java
@@ -1,4 +1,4 @@
-package step02.domain;
+package utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
